layout/inspector/tests/test_bug462787.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=462787
-->
<head>
  <title>Test for Bug 462787</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=462787">Mozilla Bug 462787</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 462787 **/

const InspectorUtils = SpecialPowers.InspectorUtils;

function do_test() {
  try {
    InspectorUtils.getCSSStyleRules(null);
    ok(false, "expected an exception"); 
  }
  catch(e) {
    is(e.name, "TypeError", "got the expected exception");
  }

  try {
    InspectorUtils.getRuleLine(null);
    ok(false, "expected an exception"); 
  }
  catch(e) {
    is(e.name, "TypeError", "got the expected exception");
  }

  try {
    InspectorUtils.isIgnorableWhitespace(null);
    ok(false, "expected an exception");
  }
  catch(e) {
    is(e.name, "TypeError", "got the expected exception");
  }

  try {
    InspectorUtils.getParentForNode(null, true);
    ok(false, "expected an exception");
  }
  catch(e) {
    is(e.name, "TypeError", "got the expected exception");
  }

  try {
    InspectorUtils.getBindingURLs(null); 
    ok(false, "expected an exception"); 
  }
  catch(e) {
    is(e.name, "TypeError", "got the expected exception");
  }

  try {
    InspectorUtils.getContentState(null); 
    ok(false, "expected an exception"); 
  }
  catch(e) {
    is(e.name, "TypeError", "got the expected exception");
  }

  try {
    InspectorUtils.setContentState(null, false); 
    ok(false, "expected an exception"); 
  }
  catch(e) {
    is(e.name, "TypeError", "got the expected exception");
  }

  try {
    InspectorUtils.setContentState(document.documentElement, 3);
    ok(false, "expected an exception");
  }
  catch(e) {
    ok(true, "Threw");
  }

  try {
    InspectorUtils.removeContentState(document.documentElement, 3);
    ok(false, "expected an exception");
  }
  catch(e) {
    ok(true, "Threw");
  }

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(do_test);


</script>
</pre>
</body>
</html>
